requirements.txt
docker-py==0.7.1
#dulwich==0.9.8
GitPython==0.3.6
pecan==0.8.3
WSME==0.6.4
PyYAML==3.11
